Quarks: Don't reload webview on screen rotation
If configChanges is not specified the Activity is restarted on every
screen rotation, causing the WebView reload the current page. This is
quite incovenient, especially with webpages that rely on javascript
to update the content. The layout already works independently on the
screen configuration and the webview adjusts its content automatically,
so configChanges is all we need.

Change-Id: I8fd99bc708ef819527412f4f30a779aea9b18901
Signed-off-by: Joe Maples <joe@frap129.org>