Quarks: Don't claim to be a simple WebView
If the user agent string includes "wv", then some websites indentify
the browser as a simple WebView and stop working properly (e.g. Google
OAuth 2.0 can't be used). Fix this removing "wv" from the user agent.
Note that this commit also changes the way the desktop user agent
is generated so that a single regular expression is used.

REGRESSION-519

Change-Id: Id138e5b38c6de7476084d7717c59f11ebe7d8178
Signed-off-by: Joe Maples <joe@frap129.org>